% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root_it.R
\name{root_it}
\alias{root_it}
\title{Calculate the nth root of a numeric value}
\usage{
root_it(x, power, replace_NA = 0)
}
\arguments{
\item{x}{Numeric value}

\item{power}{The n value, e.g. 2 for square root}

\item{replace_NA}{A character vector specifying what to replace NA values with. Defaults to 0.}
}
\description{
Calculate the nth root of a numeric value
}
\examples{
root_it(100, 2)
root_it(1024, 10)
root_it(81, 4)
}
