% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_vm_loadings.R
\name{run_vm_loadings}
\alias{run_vm_loadings}
\title{Run a varimax-rotated principal component analysis and extract loadings}
\usage{
run_vm_loadings(data, nfactors)
}
\arguments{
\item{data}{Data containing only the numeric input variables.}

\item{nfactors}{Number of vectors to be passed through.}
}
\description{
Uses `psych::principal()` and varimax rotation by default.
}
\examples{
\dontrun{ 
library(tidyverse)
iris \%>\% select(-Species) \%>\% run_loadings(nfactors = 2)
}
}
